HeroBlog refactoring
Why?
* remove backgound image from heroblog component
* move layout to components folder
* removes unused images from heroblog folger
* remove navbar to layout component
